requirements.txt
requests
loguru
paramiko
pwntools